% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{my_knn_cv}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{Input data frame.}

\item{cl}{True class value of your training data.}

\item{k_nn}{Integer representing the number of neighbors}

\item{k_cv}{Integer representing the number of folds}
}
\value{
A list with elements:\cr
\itemize{
\item \code{class} A vector of the predicted class Y^i for
all observations\cr
\item \code{cv_err} A numeric with the cross-validation
misclassification error
}
}
\description{
Uses k-Nearest Neighbors Cross-Validation method to predict
a output class using covariates.
}
\examples{
train <- iris[,1:4]
cl <- iris$Species
my_knn_cv(train, cl, 1, 5)

}
\keyword{Neighbors}
\keyword{k-Nearest}
