% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{my_lm}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{A formula class object.}

\item{data}{Input data frame.x A numeric vector of data.}
}
\value{
a table with rows for each coefficient and columns
for "Estimate", "Std. Error", "t value", and "p value"
}
\description{
Fits a linear model in R.
}
\examples{
data("mtcars")
my_lm(formula = mpg ~ hp + wt, data = mtcars)

}
\keyword{linear}
\keyword{model}
