% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{my_rf_cv}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{number of folds.}
}
\value{
A numeric with the cross-validation error
}
\description{
Uses Random Forest Cross-Validation method to predict
a output class using covariates.
}
\examples{
my_rf_cv(5)

}
\keyword{Forest}
\keyword{Random}
