% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t_test.R
\name{my_t_test}
\alias{my_t_test}
\title{my_t_test}
\usage{
my_t_test(x, alternative, mu)
}
\arguments{
\item{x}{A numeric vector of data.}

\item{alternative}{A character string specifying the alternative
hypothesis (only accept "two.sided", "less", or "greater").}

\item{mu}{A number indicating the null hypothesis value of the
mean.}
}
\value{
A list with elements:\cr
\itemize{
\item \code{test_stat} The numeric test statistic\cr
\item \code{df} The degrees of freedom\cr
\item \code{alternative} The value of the parameter \code{alternative}\cr
\item \code{p_val} The numeric p-value
}
}
\description{
Performs a one sample t-test on data.
}
\examples{
x <- rnorm(10, mean = 0, sd = 1)
my_t_test(x, alternative="two.sided", mu=0)

}
\keyword{p_value}
